change the fitted model
Adds notebook 02a which changes the offsets to relate to experiment
arrays, not individual genes.

This required changes to notebook 01 (to index on repXtrt), and the
datasets that are produced by notebook 01. Changes to functions in
the tools.py file were also required.

notebook 02a will be kept in parallel with notebook 02 until it
runs completely without modification.