modify prior distributions for phi
The Normal assumption for mean of phi is changed to Cauchy, and the
expected scale value reduced to 0.2. This results in plausible bias
values being seen for some batches (as we would have expected) and,
to offset some concerns about confirmation bias with this change,
the Rhat values now all reach 1.0 (they did not before!) indicating
that the fit converges well.